src/lib/meme-scraper/db_utils.py
import os
import psycopg2
import uuid  # Add for generating fake IDs
from typing import Optional

def get_db_connection():
    """
    Establishes a connection to the Supabase PostgreSQL database
    using the connection string from the environment variables.
    
    Note: For Chunk 1, this function will simulate a connection
    since direct database connections may not be allowed from outside Supabase.
    """
    # Check if the environment variable is set
    db_url = os.environ.get("SUPABASE_DB_URL")
    if not db_url:
        print("Error: SUPABASE_DB_URL environment variable not set.")
        raise ValueError("SUPABASE_DB_URL environment variable not set")
    
    print(f"Would connect to database using URL: {db_url}")
    
    # For Chunk 1 testing, we'll simulate a successful connection
    print("CHUNK 1 SIMULATION: Successfully connected to database.")
    return "SIMULATED_CONNECTION"

def insert_initial_reel(instagram_url: str) -> Optional[str]:
    """
    Simulates inserting a new record into the unprocessed_templates table.
    
    Args:
        instagram_url: The URL of the Instagram Reel.
        
    Returns:
        A fake UUID for Chunk 1 testing purposes.
    """
    # Generate a fake UUID for testing
    fake_reel_id = str(uuid.uuid4())
    
    # Instead of actually connecting to the database, we'll simulate it
    print(f"CHUNK 1 SIMULATION: Would insert URL {instagram_url} into unprocessed_templates table")
    print(f"CHUNK 1 SIMULATION: Generated fake ID: {fake_reel_id}")
    
    return fake_reel_id

# For future Chunks, we'll modify this file to include:
# 1. Proper database connection using connection pooling or other Supabase-approved methods
# 2. Actual inserts into the unprocessed_templates table
# 3. Functions to update the status, error_message, and other fields 