src/lib/meme-scraper/requirements.txt
yt-dlp>=2023.0.0
opencv-contrib-python>=4.7.0.0
pytesseract>=0.3.10
Pillow>=9.5.0
psycopg2-binary>=2.9.10
httpx==0.24.1
python-magic==0.4.27